continue registering rental details during rental proccess (extras listing need to be fixed)

+++ app/models/rental.py
@@ -32,7 +32,7 @@ class Extra(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(100), unique=True, nullable=False)
     description = db.Column(db.Text, nullable=False)
-    daily_price = db.Column(db.Float, nullable=False)
+    daily_price = db.Column(db.Numeric(10, 2), nullable=False)
     selectable_quantity = db.Column(db.Boolean, nullable=False)
 
     rentals = db.relationship("RentalExtra", back_populates="extra")
